change 1ボットにつき1戦略になるように変更

+++ src/Domain/Settings/StrategySetting.cs
@@ -4,5 +4,5 @@ public record class StrategySetting
 {
     public StrategyKind Kind { get; set; }
     public Timeframe Timeframe { get; set; }
-    public required IEnumerable<int> Parameters { get; set; }
+    public required IEnumerable<decimal> Parameters { get; set; }
 }
